+++ wayland-client/CHANGELOG.md
@@ -7,6 +7,8 @@
 - `Connection::null_id()` has been removed, instead use `ObjectId::null()`.
 - `EventQueue::sync_roundtrip()` has been renamed to `EventQueue::roundtrip()`.
 - Module `globals` has been removed as the abstractions it provide are not deemed useful.
+- The trait `DelegateDispatch` as been removed, its functionnality being fused into a more generic
+  version of the `Dispatch` trait.
 
 ## 0.30.0-beta.6